chore: rename variant to profile
